116[merge]

+++ Integrated_project_SC3_backend/src/main/java/sit/int221/sc3_server/DTO/Authentication/AuthUserDetail.java
@@ -7,11 +7,11 @@ import java.util.ArrayList;
 import java.util.Collection;
 @Getter
 public class AuthUserDetail extends org.springframework.security.core.userdetails.User{
-    private Long id;
-    public AuthUserDetail(Long id, String username, String password) {
+    private Integer id;
+    public AuthUserDetail(Integer id, String username, String password) {
         this(id, username, password,new ArrayList<GrantedAuthority>());
     }
-    public AuthUserDetail(Long id, String username, String password
+    public AuthUserDetail(Integer id, String username, String password
             , Collection<? extends GrantedAuthority> authorities) {
         super(username, password, authorities);
         this.id = id;
